refactor(approval): share authority snapshot sealing
